requirements.txt
azure-cognitiveservices-speech==1.35.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
scipy==1.12.0
urllib3==2.2.1
